20581 improve external type resolvers
Abstract tracking data and parsing logic from mapping layer
