+++ dina-base-api/src/main/java/ca/gc/aafc/dina/mapper/DinaMappingLayer.java
@@ -1,71 +1,38 @@
 package ca.gc.aafc.dina.mapper;
 
 import ca.gc.aafc.dina.dto.ExternalRelationDto;
-import ca.gc.aafc.dina.dto.RelatedEntity;
-import ca.gc.aafc.dina.repository.meta.JsonApiExternalRelation;
 import ca.gc.aafc.dina.service.DinaService;
 import io.crnk.core.engine.internal.utils.PropertyUtils;
 import io.crnk.core.queryspec.QuerySpec;
 import io.crnk.core.resource.annotations.JsonApiId;
-import io.crnk.core.resource.annotations.JsonApiRelation;
 import lombok.NonNull;
 import lombok.SneakyThrows;
 import org.apache.commons.lang3.reflect.FieldUtils;
 
-import java.lang.reflect.Field;
-import java.lang.reflect.Modifier;
-import java.lang.reflect.ParameterizedType;
-import java.lang.reflect.Type;
-import java.util.AbstractMap;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
 import java.util.function.BiFunction;
-import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
 public class DinaMappingLayer<D, E> {
 
-  private final Map<Class<?>, Set<String>> resourceFieldsPerClass;
-  private final Map<Class<?>, Set<String>> entityFieldsPerClass;
-  private final Map<String, String> externalNameToTypeMap;
-  private final Map<String, Class<?>> relationNamePerClass;
-  private final Set<String> collectionBasedRelations;
   private final DinaMapper<D, E> dinaMapper;
   private final DinaService<?> dinaService;
+  private final DinaMappingRegistry registry;
 
   public DinaMappingLayer(
     @NonNull Class<D> resourceClass,
     @NonNull DinaService<?> dinaService,
     @NonNull DinaMapper<D, E> dinaMapper
   ) {
-    this.resourceFieldsPerClass =
-      parseFieldsPerClass(resourceClass, new HashMap<>(), DinaMappingLayer::isNotMappable);
-    this.entityFieldsPerClass = getFieldsPerEntity();
     this.dinaService = dinaService;
     this.dinaMapper = dinaMapper;
-    this.relationNamePerClass = FieldUtils
-      .getFieldsListWithAnnotation(resourceClass, JsonApiRelation.class).stream()
-      .filter(field -> !field.isAnnotationPresent(JsonApiExternalRelation.class))
-      .collect(Collectors.toMap(
-        Field::getName,
-        field -> DinaMapper.isCollection(field.getType()) ?
-          DinaMapper.getGenericType(field.getDeclaringClass(), field.getName()) : field.getType()));
-    this.collectionBasedRelations = FieldUtils
-      .getFieldsListWithAnnotation(resourceClass, JsonApiRelation.class)
-      .stream()
-      .filter(field -> DinaMapper.isCollection(field.getType()))
-      .map(Field::getName).collect(Collectors.toSet());
-    this.externalNameToTypeMap = FieldUtils
-      .getFieldsListWithAnnotation(resourceClass, JsonApiExternalRelation.class)
-      .stream().collect(Collectors.toMap(
-        Field::getName,
-        field -> field.getAnnotation(JsonApiExternalRelation.class).type()));
+    this.registry = new DinaMappingRegistry(resourceClass);
   }
 
   public List<D> mapEntitiesToDto(@NonNull QuerySpec query, @NonNull List<E> entities) {
@@ -73,13 +40,14 @@ public class DinaMappingLayer<D, E> {
       .map(ir -> ir.getAttributePath().get(0))
       .filter(this::isNotExternal).collect(Collectors.toSet());
 
-    Map<String, Class<?>> shallowRelationsToMap = relationNamePerClass.entrySet().stream()
+    Map<String, Class<?>> shallowRelationsToMap = registry.getRelationNamePerClass()
+      .entrySet().stream()
       .filter(relation -> relationsToMap.stream().noneMatch(relation.getKey()::equalsIgnoreCase))
       .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
 
     return entities.stream()
       .map(e -> {
-        D dto = dinaMapper.toDto(e, entityFieldsPerClass, relationsToMap);
+        D dto = dinaMapper.toDto(e, registry.getEntityFieldsPerClass(), relationsToMap);
         mapShallowRelations(e, dto, shallowRelationsToMap);
         mapExternalRelationsToDto(e, dto);
         return dto;
@@ -88,136 +56,35 @@ public class DinaMappingLayer<D, E> {
   }
 
   public <S extends D> void mapToEntity(@NonNull S dto, @NonNull E entity) {
-    dinaMapper.applyDtoToEntity(dto, entity, resourceFieldsPerClass, relationNamePerClass.keySet());
-    linkRelations(entity, relationNamePerClass);
+    dinaMapper.applyDtoToEntity(
+      dto, entity,
+      registry.getResourceFieldsPerClass(),
+      registry.getRelationNamePerClass().keySet());
+    linkRelations(entity, registry.getRelationNamePerClass());
     mapExternalRelationsToEntity(dto, entity);
   }
 
   public D mapForDelete(@NonNull E entity) {
-    return dinaMapper.toDto(entity, entityFieldsPerClass, Collections.emptySet());
+    return dinaMapper.toDto(entity, registry.getEntityFieldsPerClass(), Collections.emptySet());
   }
 
   private void mapExternalRelationsToDto(E source, D target) {
-    externalNameToTypeMap.keySet().forEach(external -> {
+    registry.getExternalNameToTypeMap().keySet().forEach(external -> {
       String id = PropertyUtils.getProperty(source, external).toString();
       PropertyUtils.setProperty(target, external,
-        ExternalRelationDto.builder().type(externalNameToTypeMap.get(external)).id(id).build());
+        ExternalRelationDto.builder()
+          .type(registry.getExternalNameToTypeMap().get(external))
+          .id(id)
+          .build());
     });
   }
 
   private void mapExternalRelationsToEntity(D source, E target) {
-    externalNameToTypeMap.keySet().forEach(external -> {
+    registry.getExternalNameToTypeMap().keySet().forEach(external -> {
       Object externalRelation = PropertyUtils.getProperty(source, external);
       PropertyUtils.setProperty(target, external,
         UUID.fromString(PropertyUtils.getProperty(externalRelation, "id").toString()));
     });
-  }
-
-  /**
-   * Transverses a given class to return a map of fields per class parsed from the given class. Used
-   * to determine the necessary classes and fields per class when mapping a java bean. Fields marked
-   * with {@link JsonApiRelation} will be treated as separate classes to map and will be transversed
-   * and mapped.
-   *
-   * @param <T>            - Type of class
-   * @param clazz          - Class to parse
-   * @param fieldsPerClass - initial map to use
-   * @param ignoreIf       - predicate to return true for fields to be removed
-   * @return a map of fields per class
-   */
-  @SneakyThrows
-  private static <T> Map<Class<?>, Set<String>> parseFieldsPerClass(
-    @NonNull Class<T> clazz,
-    @NonNull Map<Class<?>, Set<String>> fieldsPerClass,
-    @NonNull Predicate<Field> ignoreIf
-  ) {
-    if (fieldsPerClass.containsKey(clazz)) {
-      return fieldsPerClass;
-    }
-
-    List<Field> relationFields = FieldUtils.getFieldsListWithAnnotation(
-      clazz,
-      JsonApiRelation.class
-    );
-
-    List<Field> attributeFields = FieldUtils.getAllFieldsList(clazz).stream()
-      .filter(f -> !relationFields.contains(f) && !f.isSynthetic() && !ignoreIf.test(f))
-      .collect(Collectors.toList());
-
-    Set<String> fieldsToInclude = attributeFields.stream()
-      .map(Field::getName)
-      .collect(Collectors.toSet());
-
-    fieldsPerClass.put(clazz, fieldsToInclude);
-
-    parseRelations(clazz, fieldsPerClass, relationFields, ignoreIf);
-
-    return fieldsPerClass;
-  }
-
-  /**
-   * Helper method to parse the fields of a given list of relations and add them to a given map.
-   *
-   * @param <T>            - Type of class
-   * @param clazz          - class containing the relations
-   * @param fieldsPerClass - map to add to
-   * @param relationFields - relation fields to transverse
-   */
-  @SneakyThrows
-  private static <T> void parseRelations(
-    Class<T> clazz,
-    Map<Class<?>, Set<String>> fieldsPerClass,
-    List<Field> relationFields,
-    Predicate<Field> removeIf
-  ) {
-    for (Field relationField : relationFields) {
-      if (Collection.class.isAssignableFrom(relationField.getType())) {
-        ParameterizedType genericType = (ParameterizedType) clazz
-          .getDeclaredField(relationField.getName())
-          .getGenericType();
-        for (Type elementType : genericType.getActualTypeArguments()) {
-          parseFieldsPerClass((Class<?>) elementType, fieldsPerClass, removeIf);
-        }
-      } else {
-        parseFieldsPerClass(relationField.getType(), fieldsPerClass, removeIf);
-      }
-    }
-  }
-
-  /**
-   * Returns a map of fields per entity class.
-   *
-   * @return a map of fields per entity class.
-   */
-  private Map<Class<?>, Set<String>> getFieldsPerEntity() {
-    return resourceFieldsPerClass.entrySet()
-      .stream()
-      .filter(e -> getRelatedEntity(e.getKey()) != null)
-      .map(e -> new AbstractMap.SimpleEntry<>(getRelatedEntity(e.getKey()).value(), e.getValue()))
-      .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
-  }
-
-  /**
-   * Returns a Dto's related entity (Marked with {@link RelatedEntity}) or else null.
-   *
-   * @param <T>   - Class type
-   * @param clazz - Class with a related entity.
-   * @return a Dto's related entity, or else null
-   */
-  private static <T> RelatedEntity getRelatedEntity(Class<T> clazz) {
-    return clazz.getAnnotation(RelatedEntity.class);
-  }
-
-  /**
-   * Returns true if the dina repo should not map the given field. currently that means if the field
-   * is generated (Marked with {@link DerivedDtoField}) or final.
-   *
-   * @param field - field to evaluate
-   * @return - true if the dina repo should not map the given field
-   */
-  private static boolean isNotMappable(Field field) {
-    return field.isAnnotationPresent(DerivedDtoField.class) ||
-           Modifier.isFinal(field.getModifiers());
   }
 
   /**
@@ -260,7 +127,7 @@ public class DinaMappingLayer<D, E> {
     for (Map.Entry<String, Class<?>> relation : relations.entrySet()) {
       String relationName = relation.getKey();
       Class<?> relationType = relation.getValue();
-      if (collectionBasedRelations.contains(relationName)) {
+      if (registry.getCollectionBasedRelations().contains(relationName)) {
         Collection<?> relationValue = (Collection<?>) PropertyUtils.getProperty(
           source, relationName);
         if (relationValue != null) {
@@ -321,7 +188,7 @@ public class DinaMappingLayer<D, E> {
   }
 
   private boolean isNotExternal(String field) {
-    return externalNameToTypeMap.keySet().stream().noneMatch(field::equalsIgnoreCase);
+    return registry.getExternalNameToTypeMap().keySet().stream().noneMatch(field::equalsIgnoreCase);
   }
 
 }
